requirements.txt
alembic==1.16.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.43
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
